internal/test: add plan config test data
Adds four plan configs for use as test data.

The diamond plan makes use of the mux & demux cryptex types and is an
example of a multi-path encryption scheme.

The two man rule plan is an encryption scheme based on the two-man
control mechanism.

The two party plan is an example of a request-approve-unlock encryption
scheme.

The DNSSEC party plan uses OpenPGP encryption keys and multi party
m-of-n encryption. It is inspired by the DNSSEC root key.