cmd/vcrypt: add DB type
The DB reads & writes serialized Material to & from the local
filesystem. Each Material is stored in an individual file named by the
ID of the material. These files are nested inside a directory named for
the ID of the vault to which they belong. They are not encrypted on
disk at this time.